Re-structured Python/ds folder
